Update cost to have decimal type

+++ InvoiceProject/InvoiceLine.cs
@@ -5,6 +5,6 @@
         public int InvoiceLineId { get; set; }
         public string Description { get; set; }
         public int Quantity { get; set; }
-        public double Cost { get; set; }
+        public decimal Cost { get; set; }
     }
 }
